Create PUBG
Arun plays PUBG a lot. One might say he is addicted to it.
He asks you to create a simple CLI-based program to show the details of a user in a high score table.
The details about the program are given below -
1. A user will have the following details
- Player Name - ex Arun (String)
- User Name - ex AKArun124 (String)
- Level - GOLD/SILVER/BRONZE (String)
- High Score (int)
There can be n number of players in a game.